Upstream sync. [MPlayer r30972 / FFmpeg r22705] Backported code from WiiMC. Regenerated configuration, disabled internal decoders. Added initial surround mixing. Added support for mono audio output. Took care of a significant bottleneck in the video driver. Moved textures to MEM2. Fixed crash when stopping while using libdvdnav. Fixed crash when stopping while a scaled DVD subtitle is displayed. Other minor fixes and changes.

+++ trunk/mplayer/mp_core.h
@@ -103,7 +103,7 @@ typedef struct MPContext {
     // by the audio CPU usage meter.
     double delay;
 
-    float begin_skip; ///< start time of the current skip while on edlout mode
+    double begin_skip; ///< start time of the current skip while on edlout mode
     // audio is muted if either EDL or user activates mute
     short edl_muted; ///< Stores whether EDL is currently in muted mode.
     short user_muted; ///< Stores whether user wanted muted mode.
